// apps/backend/src/enums/investigation.enum.ts
export enum InvestigationSubjectType {
  WALLET = 'wallet',
  CONTRACT = 'contract',
}

export enum InvestigationEntityType {
  CONTRACT = 'contract',
  WALLET = 'wallet',
  PROTOCOL = 'protocol',
  FUNDING_SOURCE = 'funding-source',
  COUNTERPARTY = 'counterparty',
  FLAGGED_SERVICE = 'flagged-service',
}